+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 From 2019 onwards, all notable changes to tarpaulin will be documented in this
 file.
 
+## [Unreleased]
+### Changed
+- No longer add `+nightly` if cargo version is already nightly
+
 ## [0.27.3] 2024-01-13
 ### Changed
 - Add line reports and make file name float in HTML report